Cleaned-up a bit of the code, exported postman as a back-up, renamed main class
