+++ src/test/java/spring/graphql/rest/nonoptimized/support/SpreadsheetsAPI.java
@@ -1,4 +1,4 @@
-package spring.graphql.rest.nonoptimized;
+package spring.graphql.rest.nonoptimized.support;
 
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
