+++ src/test/java/spring/graphql/rest/nonoptimized/support/SpreadsheetsRequest.java
@@ -1,4 +1,4 @@
-package spring.graphql.rest.nonoptimized;
+package spring.graphql.rest.nonoptimized.support;
 
 import com.google.api.services.sheets.v4.model.Request;
